+++ setup.py
@@ -70,9 +70,11 @@ setup(
     },
     entry_points={
         "console_scripts": [
-            "pip=pip._internal.main:main",
-            "pip{}=pip._internal.main:main".format(sys.version_info[0]),
-            "pip{}.{}=pip._internal.main:main".format(*sys.version_info[:2]),
+            "pip=pip._internal.cli.main:main",
+            "pip{}=pip._internal.cli.main:main".format(sys.version_info[0]),
+            "pip{}.{}=pip._internal.cli.main:main".format(
+                *sys.version_info[:2]
+            ),
         ],
     },
